Rename AvatarStackComponent to Beta::AvatarStack
