+++ CHANGELOG.md
@@ -30,12 +30,6 @@ The category for changes related to documentation, testing and tooling. Also, fo
 
 ## main
 
-### Breaking changes
-
-* Require tab with panels to have `panel_id` so `aria-controls` can be set.
-
-    *Kate Higa*
-
 ### Updates
 
 * Update the `Truncate` component to accept `:strong` as a tag.
@@ -43,6 +37,17 @@ The category for changes related to documentation, testing and tooling. Also, fo
     *Amélia Chavot*
 
 * Improve `Primer::Classify::Utilities.classes_to_hash` performance.
+
+    *Manuel Puyol*
+
+### Breaking changes
+
+* Require tab with panels to have `panel_id` so `aria-controls` can be set.
+
+    *Kate Higa*
+
+* Renames:
+  * `Primer::AvatarStackComponent` to `Primer::Beta::AvatarStack`.
 
     *Manuel Puyol*
