+++ app/components/primer/beta/avatar_stack.html.erb
