+++ static/statuses.json
@@ -1,12 +1,12 @@
 {
   "Primer::Alpha::ButtonMarketing": "alpha",
-  "Primer::AvatarStackComponent": "beta",
   "Primer::BaseButton": "beta",
   "Primer::BaseComponent": "beta",
   "Primer::Beta::AutoComplete": "beta",
   "Primer::Beta::AutoComplete::Input": "alpha",
   "Primer::Beta::AutoComplete::Item": "beta",
   "Primer::Beta::Avatar": "beta",
+  "Primer::Beta::AvatarStack": "beta",
   "Primer::Beta::Text": "beta",
   "Primer::BlankslateComponent": "beta",
   "Primer::BorderBoxComponent": "beta",
